+++ src/app/static/src/tests/components/header/fileMenu.test.ts
@@ -103,7 +103,7 @@ describe("File menu", () => {
         expect(wrapper.find(".dropdown-menu").classes()).toStrictEqual(["dropdown-menu", "show"]);
         let link = wrapper.findAll(".dropdown-item").at(0);
         link.trigger("mousedown");
-        expectTranslated(link, "Save", "Sauvegarder", store as any);
+        expectTranslated(link, "Save", "Sauvegarder", "Guardar", store as any);
 
         const hiddenLink = wrapper.find({ref: "save"});
         expect(hiddenLink.attributes("href")).toBe("http://localhost#1234");
@@ -148,7 +148,7 @@ describe("File menu", () => {
         wrapper.find(".dropdown-toggle").trigger("click");
         expect(wrapper.find(".dropdown-menu").classes()).toStrictEqual(["dropdown-menu", "show"]);
         const link = wrapper.findAll(".dropdown-item").at(1);
-        expectTranslated(link, "Load", "Charger", store as any);
+        expectTranslated(link, "Load", "Charger", "Carregar", store as any);
 
         const input = wrapper.find("input").element as HTMLInputElement;
         input.addEventListener("click", function () {
@@ -232,7 +232,7 @@ describe("File menu", () => {
 
         const modal = wrapper.find(Modal);
         modal.find(".btn").trigger("click");
-        expectTranslated(modal.find(".btn"), "OK", "OK", store as any);
+        expectTranslated(modal.find(".btn"), "OK", "OK", "OK", store as any);
         expect(clearErrorMock.mock.calls.length).toBe(1);
     });